studentsite/requirements/base.txt
django==1.10
djangorestframework==3.4.3
Markdown==2.6.6
